Rename video data header decoding operator
